+++ src/id8_i/plans/nexus_acq_rigaku_zdt_fly.py
@@ -10,7 +10,7 @@ from bluesky import plans as bp
 
 from ..utils.dm_util import dm_run_job
 from ..utils.dm_util import dm_setup
-from .nexus_utils import create_nexus_format_metadata
+from ..utils.nexus_utils import create_nexus_format_metadata
 from .sample_info_unpack import gen_folder_prefix
 from .sample_info_unpack import mesh_grid_move
 from .shutter_logic import blockbeam
